MASSIVE commit to restructure app to a RESTful style
This commit is way too large, but since I'm the only developer, it doesn't matter
that much. This commit restructures the internals and externals of the application
to be more RESTful. There are no longer any silly DirectionID or FullRoute objects,
only Agency, Route, Direction, Stop, and Prediction objects. The JSON API reflects
this change, as well.

The site now supports all agencies that Nextbus supports, thanks to discovering
the undocumented agencyList command on the Nextbus API. Hooray!

In addition, I've worked on unit testing a bit, revamping it to use the testbed
framework that comes with the App Engine SDK, and writing a few new tests.
I also had to add a lot of new XML files to the nextbus_api urlfetch mocks
directory, so that I could write unit tests for the JSON API.

The new listener form is now totally broken, since it relies on the API being in the
old format, which of course it no longer is. Fortunately, these changes lay the
groundwork for rewriting the new listener form using BackboneJS and
Backbone-relational, which will be more maintainable and extensible.

Lastly, I added some features to the flask-cache plugin, exposing the cache timeout,
cache key, and original uncached function as attributes on the decorated function.
I plan on pushing these updates back to the flask-cache project, as well.